Enh/windsor body ddes

+++ WindsorBody/assets/scripts/concatenateDomain.sh
@@ -1,4 +1,4 @@
-#!/bin/bash -l
+#!/bin/bash
 # This is a shell script to concatenate the windsor body case n times at the right hand side boundary.
 # It relies on OpenFOAM (v2206) utilities to manipulate the domain and fields.
 #
@@ -55,6 +55,12 @@ else
     echo "Concatenate/map fields of time $timeDir"
 fi
 
+if ! $concatFields && [[ -d ${timeDir} ]]; then
+    echo "Warning ${timeDir} present, and concatFields not set" 
+    echo "temporarily moving  ${timeDir} to ${timeDir}.bck"
+    mv $timeDir $timeDir.bck
+fi
+
 echo
 logDir="logFiles"
 mkdir -p $logDir
@@ -63,6 +69,8 @@ cd ..
 # --- create temporary directories with translated domain
 for ((iCopy=1; iCopy<=$nCaseCopies; iCopy++))
 do
+    echo "copy domain $iCopy of $nCaseCopies"
+
     iTmpDir=tmpDir$(( $iCopy - 1 ))
     cp -a $runDir $iTmpDir/
     cd $iTmpDir
@@ -78,6 +86,7 @@ do
     else
         # --- translate Mesh and flowfield in y-direction
         transVec="(0 $( bc -l <<<"($iCopy-1)*1.92" ) 0)"
+    echo "call transformPoints with $transVec in $PWD, placing logs in $logDir"
         transformPoints -rotateFields -translate "$transVec" > $logDir/01_transformpoints_$(($iCopy-1)).log 2>&1 || exit 1
         # --- rename all patches, except the right boundary, to have unique patch names for later apllication of mapFieldsPar utility
         # change patch names in boundaryFields of constant directory
@@ -102,6 +111,7 @@ do
     fi
     cd ..
 done 
+echo "Done translating all $nCaseCopies domains"
 
 # --- Copy and Merge mesh
 cd $runDir
@@ -110,6 +120,7 @@ if $concatFields; then
     rm -r $timeDir/*
 fi
 
+echo "Start merging all $nCaseCopies domains"
 for ((iCopy=2; iCopy<=$nCaseCopies; iCopy++))
 do
     iTmpDir="tmpDir$(($iCopy-1))"
@@ -117,16 +128,21 @@ do
     # change patch names in boundaryFields of constant directory
     sed -i "s/\bCFDWT_Right\b/CFDWT_Right_intern/g" ./constant/polyMesh/boundary
     # Merge "original" mesh with translated mesh
-    echo "mergeMeshse"
+    echo "call mergeMeshes with $overwriteFlag in $PWD, placing logs in $logDir"
     mergeMeshes $overwriteFlag . ../$iTmpDir/ >> $logDir/02_mergeMeshes.log 2>&1 || exit 1
     # stitch the merged mesh on internal patch faces
+    echo "call stitchMesh with $overwriteFlag in $PWD, placing logs in $logDir"
     stitchMesh $overwriteFlag CFDWT_Right_intern CFDWT_Left_intern >> $logDir/03_stitchMesh.log 2>&1 || exit 1
     # remove empty internal patches
+    echo "call stitchMesh with $overwriteFlag in $PWD, placing logs in $logDir"
     createPatch $overwriteFlag -dict ./system/createPatchDict.deleteIntern >> $logDir/04_createPatch.deleteIntern.log 2>&1 || exit 1
     rm ./$timeDir/meshPhi || true
+    echo "Done adding $iCopy domain"
 done
+echo "Done merging all $nCaseCopies domains"
 
 if $concatFields; then
+    echo "Start concatenating fields"
     cp -a 0.orig/* $timeDir/
     cd $timeDir/
     cp U U_0
@@ -153,9 +169,23 @@ if $concatFields; then
             sed -i "s/_2/_$iCopy/g" system/mapFieldsDict
         fi
 
+    echo "call mapFieldsPar in $PWD, placing logs in $logDir"
         # mapFieldsPar -sourceTime 'latestTime' -mapMethod direct -patchMapMethod nearestFaceAMI ../tmpDir0/ > $logDir/11_mapFieldsPar.log 2>&1 || exit 1 # Error, if OF-v2206 is loaded with gnu/11
         mapFieldsPar -sourceTime 'latestTime' -mapMethod direct -fields '("p.*" "U.*" "k.*" "nut.*" "nuTilda.*" "omega.*")' -targetRegion region0 ../$iTmpDir/ >> $logDir/06_mapFieldsPar.log 2>&1 || exit 1 # Error, if OF-v2206 is loaded with gnu/11
     done
+    echo "Done concatenating fields"
 fi
 
-rm -r ../tmpDir*
+if ! $concatFields && [[ -d ${timeDir}.bck ]]; then
+    echo "Restoring time directory ${timeDir}" 
+    mv $timeDir.bck $timeDir
+fi
+
+for ((iCopy=0; iCopy<=$nCaseCopies; iCopy++))
+do
+    if [[ -d ../tmpDir$iCopy ]]; then
+        echo "cleaning up tmpDir directory tmpDir$iCopy" 
+        rm -r ../tmpDir*
+    fi 
+done 
+
